+++ Project-Chapeau herkansers 3/Afrekenen.cs
@@ -35,7 +35,7 @@ namespace Project_Chapeau_herkansers_3
         public void RefreshBillItems() {
             List<BesteldeItem> besteldeItems = serviceBI.GetItemsFromBestelling(bestelling.Id);
             billListView.Items.Clear();
-            decimal total = 0.00M;
+            double total = 0.00;
             foreach (BesteldeItem b in besteldeItems)
             {
